refactor: 움직임 결과 테스트코드 패키지 구조 변경
